refactor(*): extract http-routing-testing component
Move all testing related class of the
http-routing package to a dedicated
package.

See: #53
