+++ .github/workflows/split_tests.yaml
@@ -159,6 +159,6 @@ jobs:
           cat ./tests/.env.testing.dist >> $GITHUB_ENV
           echo ${{github.env}}
 
-      - name: 'Run codeception tests'
+      - name: 'Run codeception tests of package ${{ matrix.packages.full_name }}'
         run: vendor/bin/codecept run wordpress
         working-directory: ${{ matrix.packages.absolute_path }}
