+++ composer.json
@@ -100,6 +100,7 @@
             "Snicco\\Component\\Eloquent\\": "src/Snicco/Component/eloquent/src",
             "Snicco\\Component\\EventDispatcher\\": "src/Snicco/Component/event-dispatcher/src",
             "Snicco\\Component\\HttpRouting\\": "src/Snicco/Component/http-routing/src",
+            "Snicco\\Component\\HttpRouting\\Testing\\": "src/Snicco/Component/http-routing-testing/src",
             "Snicco\\Component\\Kernel\\": "src/Snicco/Component/kernel/src",
             "Snicco\\Component\\Psr7ErrorHandler\\": "src/Snicco/Component/psr7-error-handler/src",
             "Snicco\\Component\\Session\\": "src/Snicco/Component/session/src",
@@ -154,6 +155,7 @@
             "Snicco\\Component\\Eloquent\\Tests\\": "src/Snicco/Component/eloquent/tests",
             "Snicco\\Component\\EventDispatcher\\Tests\\": "src/Snicco/Component/event-dispatcher/tests",
             "Snicco\\Component\\HttpRouting\\Testing\\": "src/Snicco/Component/http-routing/testing",
+            "Snicco\\Component\\HttpRouting\\Testing\\Tests\\": "src/Snicco/Component/http-routing-testing/tests",
             "Snicco\\Component\\HttpRouting\\Tests\\": "src/Snicco/Component/http-routing/tests",
             "Snicco\\Component\\Kernel\\Tests\\": "src/Snicco/Component/kernel/tests",
             "Snicco\\Component\\Psr7ErrorHandler\\Tests\\": "src/Snicco/Component/psr7-error-handler/tests",
@@ -264,6 +266,7 @@
         "sniccowp/guests-only-middleware": "self.version",
         "sniccowp/http-routing": "self.version",
         "sniccowp/http-routing-bundle": "self.version",
+        "sniccowp/http-routing-testing": "self.version",
         "sniccowp/https-only-middleware": "self.version",
         "sniccowp/illuminate-container-bridge": "self.version",
         "sniccowp/kernel": "self.version",
