+++ src/Snicco/Component/http-routing-testing/.gitattributes
@@ -1,0 +1,5 @@
+.gitattributes export-ignore
+.gitignore export-ignore
+phpunit.dist.xml export-ignore
+*.md export-ignore
+/tests export-ignore
