+++ src/Snicco/Component/http-routing-testing/.github/workflows/auto_close_issue.yaml
@@ -1,0 +1,19 @@
+name: Automatically close PR's
+
+on:
+  pull_request_target:
+    types: [ opened ]
+
+jobs:
+  run:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: superbrothers/close-pull-request@v3
+        with:
+          # Optional. Post a issue comment just before closing a pull request.
+          comment: |
+            Hi, thank you for your contribution.
+            Unfortunately, this repository is read-only. It's a split from our main development repository.
+            We'd like to kindly ask you to move the contribution there - https://github.com/sniccowp/sniccowp.
+            We'll check it, review it and give you feed back right way.
+            Thank you.
