+++ src/Snicco/Component/http-routing-testing/.gitignore
@@ -1,0 +1,3 @@
+/.phpunit.cache/
+/vendor
+composer.lock
